Change: SessionCore to abstract.
