+++ SynicSugar/Assets/SynicSugar/Runtime/Core/EOS/EOSCoreFactory.cs
@@ -1,7 +1,9 @@
 using SynicSugar.Auth;
 using SynicSugar.Auth.Base;
+using SynicSugar.Base;
 using SynicSugar.MatchMake;
 using SynicSugar.MatchMake.Base;
+using SynicSugar.P2P;
 namespace SynicSugar {
     internal sealed class EOSCoreFactory : SynicSugarCoreFactory {
         AuthenticationCore auth;
@@ -13,6 +15,9 @@ namespace SynicSugar {
         }
         public override MatchmakingCore GetMatchmakingCore(uint maxSearch){
             return new EOSMatchmaking(maxSearch);
+        }
+        public override SessionCore GetSessionCore(){
+            return new EOSSessionManager();
         }
     }
 }
