+++ SynicSugar/Assets/SynicSugar/Runtime/p2p/EOS/EOSSessionManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f336c22e48d874b7fb55073eea4c00e0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
